+++ csfunctions/events/__init__.py
@@ -1,4 +1,4 @@
-from typing import Annotated, Union
+from typing import Annotated
 
 from pydantic import Field
 
@@ -31,50 +31,48 @@ from .part_released import PartReleasedData, PartReleasedEvent
 from .workflow_task_trigger import WorkflowTaskTriggerEvent, WorkflowTaskTriggerEventData
 
 Event = Annotated[
-    Union[
-        DocumentReleasedEvent,
-        DocumentReleaseCheckEvent,
-        DocumentFieldCalculationEvent,
-        PartReleasedEvent,
-        PartReleaseCheckEvent,
-        PartFieldCalculationEvent,
-        FieldValueCalculationEvent,
-        DummyEvent,
-        EngineeringChangeReleasedEvent,
-        EngineeringChangeReleaseCheckEvent,
-        EngineeringChangeStatusChangedEvent,
-        EngineeringChangeStatusChangeCheckEvent,
-        WorkflowTaskTriggerEvent,
-        DocumentCreateCheckEvent,
-        DocumentModifyCheckEvent,
-        PartCreateCheckEvent,
-        PartModifyCheckEvent,
-        CustomOperationDocumentEvent,
-        CustomOperationPartEvent,
-    ],
+    DocumentReleasedEvent
+    | DocumentReleaseCheckEvent
+    | DocumentFieldCalculationEvent
+    | PartReleasedEvent
+    | PartReleaseCheckEvent
+    | PartFieldCalculationEvent
+    | FieldValueCalculationEvent
+    | DummyEvent
+    | EngineeringChangeReleasedEvent
+    | EngineeringChangeReleaseCheckEvent
+    | EngineeringChangeStatusChangedEvent
+    | EngineeringChangeStatusChangeCheckEvent
+    | WorkflowTaskTriggerEvent
+    | DocumentCreateCheckEvent
+    | DocumentModifyCheckEvent
+    | PartCreateCheckEvent
+    | PartModifyCheckEvent
+    | CustomOperationDocumentEvent
+    | CustomOperationPartEvent,
     Field(discriminator="name"),
 ]
-EventData = Union[
-    DocumentReleasedData,
-    DocumentReleaseCheckData,
-    DocumentFieldCalculationData,
-    PartReleasedData,
-    PartReleaseCheckData,
-    PartFieldCalculationData,
-    FieldValueCalculationData,
-    DummyEventData,
-    EngineeringChangeReleasedData,
-    EngineeringChangeReleaseCheckData,
-    EngineeringChangeStatusChangedData,
-    EngineeringChangeStatusChangeCheckData,
-    WorkflowTaskTriggerEventData,
-    DocumentCreateCheckData,
-    DocumentModifyCheckData,
-    PartCreateCheckData,
-    PartModifyCheckData,
-    CustomOperationDocumentData,
-    CustomOperationPartData,
-]
+EventData = (
+    DocumentReleasedData
+    | DocumentReleaseCheckData
+    | DocumentFieldCalculationData
+    | PartReleasedData
+    | PartReleaseCheckData
+    | PartFieldCalculationData
+    | FieldValueCalculationData
+    | DummyEventData
+    | EngineeringChangeReleasedData
+    | EngineeringChangeReleaseCheckData
+    | EngineeringChangeStatusChangedData
+    | EngineeringChangeStatusChangeCheckData
+    | WorkflowTaskTriggerEventData
+    | DocumentCreateCheckData
+    | DocumentModifyCheckData
+    | PartCreateCheckData
+    | PartModifyCheckData
+    | CustomOperationDocumentData
+    | CustomOperationPartData
+)
 
 __all__ = [
     "DocumentReleasedEvent",